if this doesn't work im deleting my codE

+++ TeamCode/src/main/java/org/firstinspires/ftc/team9202hme/BananaFruit.java
@@ -176,7 +176,7 @@ public class BananaFruit {
         return String.format(Locale.getDefault(), "%.1f", AngleUnit.DEGREES.normalize(degrees));
     }
 
-    public float getHeading(){
-        return angles.firstAngle;
+    public int getHeading(){
+        return (int)angles.firstAngle;
     }
 }
